refactor: generalize core client orchestration and extract base factory
