+++ src/client/factory.ts
@@ -1,18 +1,13 @@
-import type { Client as LibsqlClient } from "@libsql/client";
-import { Store } from "n3";
 import { QueryEngine } from "@comunica/query-sparql-rdfjs-lite";
-import { RecursiveCharacterTextSplitter } from "@langchain/textsplitters";
+import { Store } from "n3";
 
 import { Client, type ClientOptions } from "./client.ts";
 import type { ClientInterface } from "./interface.ts";
 import { createIndexedStore } from "./quad-store/indexed-store.ts";
 import { RdfjsQuadStore } from "./quad-store/quad-store.ts";
 import { ComunicaSparqlEngine } from "./sparql-engine/sparql-engine.ts";
-import { LibsqlSearchIndex } from "./search-index/providers/libsql/libsql-search-index.ts";
-import { LibsqlIndexSync } from "./search-index/providers/libsql/libsql-index-sync.ts";
-import { QuadChunker } from "./search-index/chunking/quad-chunker.ts";
-import { hydrateStoreFromLibsql } from "./search-index/providers/libsql/libsql-loader.ts";
-
+import type { SearchIndexInterface } from "./search-index/mod.ts";
+import type { Patch } from "./quad-store/patch.ts";
 import type {
   ImportRequest,
   ImportResponse,
@@ -21,22 +16,17 @@ import type {
   SparqlRequest,
   SparqlResponse,
 } from "./sparql-engine/mod.ts";
-import type { EmbeddingService } from "./search-index/providers/libsql/libsql-search-index.ts";
-import {
-  makeLibsqlChunksFtsTable,
-  makeLibsqlChunksIndex,
-  makeLibsqlChunksQuadIdIndex,
-  makeLibsqlChunksTable,
-  makeLibsqlChunksTriggers,
-  makeLibsqlQuadsTable,
-} from "./search-index/providers/libsql/statements.ts";
 
-const clientCache = new WeakMap<LibsqlClient, ClientInterface>();
+const queryEngine = new QueryEngine();
 
-class AutoSyncClient extends Client {
+/**
+ * SynchronizedClient is a wrapper around Client that automatically commits
+ * changes to the database after each operation.
+ */
+export class SynchronizedClient extends Client {
   constructor(
     opts: ClientOptions,
-    private flushPending: () => Promise<void>,
+    private readonly commitChanges: () => Promise<void>,
   ) {
     super(opts);
   }
@@ -45,7 +35,7 @@ class AutoSyncClient extends Client {
     request: ImportRequest,
   ): Promise<ImportResponse> {
     const res = await super.import(request);
-    await this.flushPending();
+    await this.commitChanges();
     return res;
   }
 
@@ -53,102 +43,74 @@ class AutoSyncClient extends Client {
     request: SparqlRequest,
   ): Promise<SparqlResponse> {
     const res = await super.sparql(request);
-    await this.flushPending();
+    await this.commitChanges();
     return res;
   }
 }
 
 /**
- * Configuration bundle for bootstrapping the integrated environment.
+ * Configuration bundle for generating a reactive engine deployment.
  */
-export interface CreateClientOptions {
-  /** Fully initialized LibSQL client pointing to file or in-memory target. */
-  db: LibsqlClient;
-  /** Service projected for transforming text literals into comparison vectors. */
-  embeddingService: EmbeddingService;
-  /** Optional pre-baked chunker, defaults internally if omitted. */
-  chunker?: QuadChunker;
+export interface BaseClientOptions {
+  /** 
+   * Primary semantic lookups engine powering natural language searches.
+   */
+  searchIndex: SearchIndexInterface;
+  /** 
+   * Inversion hook dispatching mutation sets to external durable storage agents.
+   */
+  sync?: (patch: Patch) => Promise<unknown>;
+  /** 
+   * Optional initializer reconstructing initial graph contents before activation.
+   */
+  hydrate?: (store: Store) => Promise<unknown>;
 }
 
 /**
- * initializeSchema synchronously checks and creates the full set of persistent tables needed.
+ * createBaseClient synthesizes the generalized orchestration machinery required to power
+ * any transactional reactive World environment. It intercepts memory mutations, cascades
+ * commitment lifecycle events, and instantiates the semantic execution graph.
  */
-async function initializeSchema(db: LibsqlClient): Promise<void> {
-  await db.execute(makeLibsqlQuadsTable());
-  await db.execute(makeLibsqlChunksTable());
-  await db.execute(makeLibsqlChunksQuadIdIndex());
-  await db.execute(makeLibsqlChunksFtsTable());
-  await db.execute(makeLibsqlChunksIndex());
-  for (const sql of makeLibsqlChunksTriggers()) {
-    await db.execute(sql);
-  }
-}
-
-/**
- * createClient handles total orchestration of the unified ecosystem.
- * It guarantees schema health, hydrates physical memory from disk, instantiates the observer bridge,
- * and returns a higher-order Client providing transactional auto-flushing.
- */
-export async function createClient(
-  options: CreateClientOptions,
+export async function createBaseClient(
+  options: BaseClientOptions,
 ): Promise<ClientInterface> {
-  const { db, embeddingService } = options;
-
-  const cached = clientCache.get(db);
-  if (cached) return cached;
-
-  // 1. Guarantee logical SQL schema existence
-  await initializeSchema(db);
-
-  // 2. Initialize and hydrate master in-memory graph
   const rawStore = new Store();
-  await hydrateStoreFromLibsql(db, rawStore);
 
-  // 3. Install dynamic Proxy Bridge tracking memory mutations
+  // Allow provider to pre-load graph dependencies.
+  if (options.hydrate) {
+    await options.hydrate(rawStore);
+  }
+
+  // Construct active monitoring bridge.
   const { store, queue } = createIndexedStore(rawStore);
 
-  // 4. Initialize dedicated IO synchronizer
-  const chunker = options.chunker ??
-    new QuadChunker({
-      splitter: new RecursiveCharacterTextSplitter({ chunkSize: 1000 }),
-    });
-
-  const synchronizer = new LibsqlIndexSync({
-    client: db,
-    embeddingService,
-    chunker,
-  });
-
-  // Inner trigger flushing the collection of mutations to disk
-  const flushPending = async () => {
+  // Centralized coordinator aggregating diff streams for upstream emission.
+  const commitChanges = async () => {
     const patches = queue.flush();
     if (patches.length === 0) return;
-    const merged = {
-      insertions: patches.flatMap((p) => p.insertions),
-      deletions: patches.flatMap((p) => p.deletions),
-    };
-    await synchronizer.sync(merged);
+    
+    if (options.sync) {
+      const merged: Patch = {
+        insertions: patches.flatMap((p) => p.insertions),
+        deletions: patches.flatMap((p) => p.deletions),
+      };
+      await options.sync(merged);
+    }
   };
 
-  // 5. Construct services sharing the Proxied observer store
+  // Attach standard core runtime facilities.
   const quadStore = new RdfjsQuadStore(store);
   const sparqlEngine = new ComunicaSparqlEngine({
     store,
-    queryEngine: new QueryEngine(),
-  });
-  const searchIndex = new LibsqlSearchIndex({
-    client: db,
-    embeddingService,
+    queryEngine,
   });
 
-  // 6. Return Higher-Order Client enforcing post-operation synchronization
+  // Assemble final high-order active composition.
   const baseOptions: ClientOptions = {
     quadStore,
     sparqlEngine,
-    searchIndex,
+    searchIndex: options.searchIndex,
   };
 
-  const client = new AutoSyncClient(baseOptions, flushPending);
-  clientCache.set(db, client);
-  return client;
+  return new SynchronizedClient(baseOptions, commitChanges);
 }
