+++ src/client/mod.ts
@@ -4,3 +4,4 @@ export * from "./quad-store/mod.ts";
 export * from "./sparql-engine/mod.ts";
 export * from "./search-index/mod.ts";
 export * from "./factory.ts";
+export * from "./search-index/providers/libsql/libsql-client.ts";
