+++ src/client/search-index/providers/libsql/mod.ts
@@ -1,1 +1,2 @@
 export * from "./libsql-search-index.ts";
+export * from "./libsql-client.ts";
